LayoutTests/platform/chromium/fast/forms/suggestion-picker/datetimelocal-suggestion-picker-key-operations.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/suggestion-picker-common.js"></script>
</head>
<body style="background-color: #bbbbbb;">
<p id="description"></p>
<div id="console"></div>
<input type=datetime-local id=datetime-local value="2012-11-22T01:00" list=suggestions>
<datalist id=suggestions>
    <option label="Now">2012-11-22T01:01</option>
    <option>2012-11-22T01:02</option>
    <option>2012-11-22T01:03</option>
    <option>2012-11-22T01:04</option>
    <option>2012-11-22T01:05</option>
    <option>2012-11-22T01:06</option>
    <option>2012-11-22T01:07</option>
    <option>2012-11-22T01:08</option>
    <option>2012-11-22T01:09</option>
    <option>2012-11-22T01:10</option>
    <option>2012-11-22T01:11</option>
    <option>2012-11-22T01:12</option>
    <option>2012-11-22T01:13</option>
    <option>2012-11-22T01:14</option>
    <option>2012-11-22T01:15</option>
    <option>2012-11-22T01:16</option>
    <option>2012-11-22T01:17</option>
    <option>2012-11-22T01:18</option>
    <option>2012-11-22T01:19</option>
    <option>2012-11-22T01:20</option>
    <option>2012-11-22T01:21</option>
    <option>2012-11-22T01:22</option>
    <option>2012-11-22T01:23</option>
    <option>2012-11-22T01:24</option>
    <option>2012-11-22T01:25</option>
    <option>2012-11-22T01:26</option>
    <option>2012-11-22T01:27</option>
    <option>2012-11-22T01:28</option>
    <option>2012-11-22T01:29</option>
    <option>2012-11-22T01:30</option>
    <option>2012-11-22T01:31</option>
    <option>2012-11-22T01:32</option>
    <option>2012-11-22T01:33</option>
    <option>2012-11-22T01:34</option>
    <option>2012-11-22T01:35</option>
    <option>2012-11-22T01:36</option>
    <option>2012-11-22T01:37</option>
    <option>2012-11-22T01:38</option>
    <option>2012-11-22T01:39</option>
    <option>2012-11-22T01:40</option>
    <option>2012-11-22T01:41</option>
    <option>2012-11-22T01:42</option>
    <option>2012-11-22T01:43</option>
    <option>2012-11-22T01:44</option>
    <option>2012-11-22T01:45</option>
    <option>2012-11-22T01:46</option>
    <option>2012-11-22T01:47</option>
    <option>2012-11-22T01:48</option>
    <option>2012-11-22T01:49</option>
    <option>2012-11-22T01:50</option>
    <option>2012-11-22T01:51</option>
    <option>2012-11-22T01:52</option>
    <option>2012-11-22T01:53</option>
    <option>2012-11-22T01:54</option>
    <option>2012-11-22T01:55</option>
    <option>2012-11-22T01:56</option>
    <option>2012-11-22T01:57</option>
    <option>2012-11-22T01:58</option>
    <option>2012-11-22T01:59</option>
    <option>2012-11-22T02:00</option>
    <option>2012-11-22T02:01</option>
    <option>2012-11-22T02:02</option>
    <option>2012-11-22T02:03</option>
    <option>2012-11-22T02:04</option>
    <option>2012-11-22T02:05</option>
    <option>2012-11-22T02:06</option>
    <option>2012-11-22T02:07</option>
</datalist>
<script>
debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

openPicker(document.getElementById('datetime-local'), test1);
shouldBeEqualToString('popupWindow.pagePopupController.toString()', '[object PagePopupController]');

function test1() {
    debug('Check that page popup exists.');
    shouldBeEqualToString('popupWindow.pagePopupController.toString()', '[object PagePopupController]');
    popupWindow.focus();

    popupWindow.removeEventListener('resize', test1, false);

    debug('Check that up/down arrow keys work.');
    shouldBeNull('highlightedEntry()');
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T01:01');
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T01:02');
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T01:01');

    debug('Check that up arrow key at top of list does nothing.');
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T01:01');
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T01:02');

    debug('Check that page up/down arrow keys work.');
    eventSender.keyDown('pageDown');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T01:40');
    eventSender.keyDown('pageDown');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T02:00');
    eventSender.keyDown('pageUp');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T01:21');
    eventSender.keyDown('pageDown');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T02:00');
    eventSender.keyDown('pageDown');
    shouldBeEqualToString('highlightedEntry()', '@openCalendarPicker');

    debug('Check that down arrow key at bottom of list does nothing.');
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('highlightedEntry()', '@openCalendarPicker');
    
    debug('Check that enter key sets the highlighted value.');
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('highlightedEntry()', '2012-11-22T02:07');
    eventSender.keyDown('\n');
    waitUntilClosing(test1AfterClosing);
}

function test1AfterClosing() {
    shouldBeNull('document.getElementById("mock-page-popup")');
    shouldBeEqualToString('document.getElementById("datetime-local").value', '2012-11-22T02:07');

    debug('Reopen popup.');
    openPicker(document.getElementById('datetime-local'));
    popupWindow.addEventListener('resize', test2, false);
}

function test2() {
    popupWindow.removeEventListener('resize', test2, false);

    popupWindow.focus();

    debug('Check that escape key closes the popup.');
    eventSender.keyDown('\x1B');
    waitUntilClosing(test2AfterClosing);
}

function test2AfterClosing() {
    shouldBeNull('document.getElementById("mock-page-popup")');
    shouldBeEqualToString('document.getElementById("datetime-local").value', '2012-11-22T02:07');

    sendKey(document.getElementById('datetime-local'), 'F4');
    if (document.getElementById('mock-page-popup'))
        testPassed('F4 opened picker.');

    finishJSTest();
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
