LayoutTests/platform/chromium/fast/forms/suggestion-picker/datetimelocal-suggestion-picker-key-operations-expected.txt

Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
Check that page popup exists.
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
Check that up/down arrow keys work.
PASS highlightedEntry() is null
PASS highlightedEntry() is "2012-11-22T01:01"
PASS highlightedEntry() is "2012-11-22T01:02"
PASS highlightedEntry() is "2012-11-22T01:01"
Check that up arrow key at top of list does nothing.
PASS highlightedEntry() is "2012-11-22T01:01"
PASS highlightedEntry() is "2012-11-22T01:02"
Check that page up/down arrow keys work.
PASS highlightedEntry() is "2012-11-22T01:40"
PASS highlightedEntry() is "2012-11-22T02:00"
PASS highlightedEntry() is "2012-11-22T01:21"
PASS highlightedEntry() is "2012-11-22T02:00"
PASS highlightedEntry() is "@openCalendarPicker"
Check that down arrow key at bottom of list does nothing.
PASS highlightedEntry() is "@openCalendarPicker"
Check that enter key sets the highlighted value.
PASS highlightedEntry() is "2012-11-22T02:07"
PASS document.getElementById("mock-page-popup") is null
PASS document.getElementById("datetime-local").value is "2012-11-22T02:07"
Reopen popup.
Check that escape key closes the popup.
PASS document.getElementById("mock-page-popup") is null
PASS document.getElementById("datetime-local").value is "2012-11-22T02:07"
PASS successfullyParsed is true

TEST COMPLETE
